requirements.txt
groq
numpy
tavily-python
xgboost
